requirements.txt
Flask>=3.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
Werkzeug>=3.0.0
python-telegram-bot>=21.0
gunicorn>=21.2.0
